Complete 16 vdo

+++ src/main/java/com/lothin/phoneshp/dto/BrandDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class BrandDTO {
-    private Long id;
-    private String name;
+    private Integer id;
+	private String name;
 }
