+++ src/main/java/com/lothin/phoneshp/dto/PaginationDTO.java
@@ -6,12 +6,12 @@ import lombok.Data;
 @Data
 @Builder
 public class PaginationDTO {
-    private Long numberOfElements;
-    private Long number;
-    private Long size;
-    private Long totalElements;
-    private Long totalPages;
-    private boolean empty;
-    private boolean first;
-    private boolean last;
+    private int numberOfElements;
+	private int number;
+	private int size;
+	private long totalElements;
+	private int totalPages;
+	private boolean empty;
+	private boolean first;
+	private boolean last;
 }
